939: Support paypal email in meal edit form also
